requirements.txt
discord.py==1.4.1
google-api-core==1.22.2
google-api-python-client==1.11.0
google-auth==1.21.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
SQLAlchemy==1.3.16
